Safely use the hash table in ddt_to_info.
Signed-off-by: George Bosilca <bosilca@icl.utk.edu>

Disable reuse for now.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>

Dont be eager to free the datatype.

A datatype has multiple associations with an arena, and the arenas are
correctly refcounted. If we release the datatype once we release one refcount
on the arenas bad things will happen. Instead only free the datatype once we
released all associated arenas.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>